+++ common/src/main/java/com/slymask3/instantblocks/block/instant/InstantEscapeLadderBlock.java
@@ -53,7 +53,7 @@ public class InstantEscapeLadderBlock extends InstantBlock implements SimpleWate
 	}
 
 	public boolean build(Level world, int x, int y, int z, Player player) {
-		Builder builder = new Builder(1,true);
+		Builder builder = new Builder(1, Direction.UP);
 
 		Block ladder = Blocks.LADDER;
 		Block torch = Blocks.TORCH;