+++ common/src/main/java/com/slymask3/instantblocks/block/instant/InstantStatueBlock.java
@@ -86,7 +86,7 @@ public class InstantStatueBlock extends InstantBlock implements EntityBlock {
 	}
 
 	public boolean build(Level world, int x, int y, int z, Player player) {
-		Builder builder = new Builder(1,true);
+		Builder builder = new Builder(2, Direction.UP);
 		
 		StatueBlockEntity blockEntity = (StatueBlockEntity)world.getBlockEntity(new BlockPos(x,y,z));
 		Direction direction = world.getBlockState(new BlockPos(x,y,z)).getValue(FACING);