+++ common/src/main/java/com/slymask3/instantblocks/block/instant/InstantTreeBlock.java
@@ -10,6 +10,7 @@ import com.slymask3.instantblocks.reference.Strings;
 import com.slymask3.instantblocks.util.ClientHelper;
 import com.slymask3.instantblocks.util.Helper;
 import net.minecraft.core.BlockPos;
+import net.minecraft.core.Direction;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.level.BlockGetter;
 import net.minecraft.world.level.Level;
@@ -51,7 +52,7 @@ public class InstantTreeBlock extends InstantBlock implements EntityBlock {
 	}
 	
 	public boolean build(Level world, int x_center, int y, int z_center, Player player) {
-		Builder builder = new Builder(1,true);
+		Builder builder = new Builder(2, Direction.UP);
 		TreeBlockEntity blockEntity = (TreeBlockEntity)world.getBlockEntity(new BlockPos(x_center,y,z_center));
 		int size = Common.CONFIG.TREE_SIZE();
 		int half = (int)Math.floor(size / 2);