+++ common/src/main/java/com/slymask3/instantblocks/util/ClientHelper.java
@@ -10,10 +10,11 @@ import net.minecraft.client.Minecraft;
 import net.minecraft.client.color.block.BlockColor;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.particles.DustParticleOptions;
+import net.minecraft.core.particles.ParticleOptions;
+import net.minecraft.core.particles.ParticleTypes;
 import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.sounds.SoundEvent;
-import net.minecraft.sounds.SoundEvents;
 import net.minecraft.sounds.SoundSource;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.level.BlockAndTintGetter;
@@ -25,48 +26,46 @@ import java.util.List;
 
 public class ClientHelper {
     public enum Screen { STATUE, HARVEST, SKYDIVE, SCHEMATIC, TREE }
-    public enum Particles { NONE, GENERATE, NO_LIQUID }
+    public enum Particles { GENERATE, NO_LIQUID, PLACE_BLOCK }
 
     public static List<ColorSet> SKYDIVE_PRESETS;
     public static List<HugeTree> HUGE_TREES;
 
-    public static void playSound(Player player, BlockPos pos, SoundEvent sound) {
-        //Common.LOG.info("sound: " + sound);
+    public static void playSound(Player player, BlockPos pos, SoundEvent sound, float volume) {
         if(sound != null) {
-            player.level.playSound(player, pos, sound, SoundSource.BLOCKS, 0.2F, 1.0F);
+            player.level.playSound(player, pos, sound, SoundSource.BLOCKS, volume, 1.0F);
         }
     }
 
-    public static void playSound(Level world, BlockPos pos, Particles particles) {
-        SoundEvent sound = switch(particles) {
-            case GENERATE -> new SoundEvent(new ResourceLocation("minecraft", Common.CONFIG.SOUND_GENERATE()));
-            case NO_LIQUID -> new SoundEvent(new ResourceLocation("minecraft", Common.CONFIG.SOUND_NO_LIQUID()));
-            default -> SoundEvents.PLAYER_LEVELUP;
-        };
-        //world.playSound(Minecraft.getInstance().player, pos, sound, SoundSource.BLOCKS,0.4F,1.0F);
-    }
-
-    public static void showParticles(Level world, BlockPos pos, Particles particles) {
-        if(Common.CONFIG.SHOW_EFFECTS()) {
-            switch(particles) {
-                case GENERATE -> {
-                    for(double i = 0; i <= 1; i = i + 0.2) {
-                        for(double n = 0; n <= 1; n = n + 0.2) {
-                            for(double v = 0; v <= 1; v = v + 0.2) {
-                                world.addParticle(DustParticleOptions.REDSTONE,pos.getX()+i, pos.getY()+v, pos.getZ()+n, 0.0D, 0.0D, 0.0D);
-                            }
+    public static void showParticles(Player player, BlockPos pos, Particles particles) {
+        Level world = player.getLevel();
+        switch(particles) {
+            case GENERATE -> {
+                for(double i = 0; i <= 1; i = i + 0.2) {
+                    for(double n = 0; n <= 1; n = n + 0.2) {
+                        for(double v = 0; v <= 1; v = v + 0.2) {
+                            addParticle(world,pos,DustParticleOptions.REDSTONE,i,v,n);
                         }
                     }
                 }
-                case NO_LIQUID -> {
-                    world.addParticle(InstantLiquidBlock.WHITE_DUST, (double)pos.getX() + 0.5D, (double)pos.getY() + 1.2D, (double)pos.getZ() + 0.5D, 0.0D, 0.0D, 0.0D);
-                    world.addParticle(InstantLiquidBlock.WHITE_DUST, (double)pos.getX() + 1.2D, (double)pos.getY() + 0.5D, (double)pos.getZ() + 0.5D, 0.0D, 0.0D, 0.0D);
-                    world.addParticle(InstantLiquidBlock.WHITE_DUST, (double)pos.getX() + 0.5D, (double)pos.getY() + 0.5D, (double)pos.getZ() + 1.2D, 0.0D, 0.0D, 0.0D);
-                    world.addParticle(InstantLiquidBlock.WHITE_DUST, (double)pos.getX() + 0.5D, (double)pos.getY() - 0.2D, (double)pos.getZ() + 0.5D, 0.0D, 0.0D, 0.0D);
-                    world.addParticle(InstantLiquidBlock.WHITE_DUST, (double)pos.getX() - 0.2D, (double)pos.getY() + 0.5D, (double)pos.getZ() + 0.5D, 0.0D, 0.0D, 0.0D);
-                    world.addParticle(InstantLiquidBlock.WHITE_DUST, (double)pos.getX() + 0.5D, (double)pos.getY() + 0.5D, (double)pos.getZ() - 0.2D, 0.0D, 0.0D, 0.0D);
-                }
+                playSound(player, pos, new SoundEvent(new ResourceLocation("minecraft", Common.CONFIG.SOUND_GENERATE())), 0.4F);
             }
+            case NO_LIQUID -> {
+                addParticle(world,pos,InstantLiquidBlock.WHITE_DUST,0.5D,1.2D,0.5D);
+                addParticle(world,pos,InstantLiquidBlock.WHITE_DUST,1.2D,0.5D,0.5D);
+                addParticle(world,pos,InstantLiquidBlock.WHITE_DUST,0.5D,0.5D,1.2D);
+                addParticle(world,pos,InstantLiquidBlock.WHITE_DUST,0.5D,-0.2D,0.5D);
+                addParticle(world,pos,InstantLiquidBlock.WHITE_DUST,-0.2D,0.5D,0.5D);
+                addParticle(world,pos,InstantLiquidBlock.WHITE_DUST,0.5D,0.5D,-0.2D);
+                playSound(player, pos, new SoundEvent(new ResourceLocation("minecraft", Common.CONFIG.SOUND_NO_LIQUID())), 0.4F);
+            }
+            case PLACE_BLOCK -> addParticle(world,pos, ParticleTypes.ENCHANT,0.5D,0.5D,0.5D);
+        }
+    }
+
+    private static void addParticle(Level world, BlockPos pos, ParticleOptions particle, double x, double y, double z) {
+        if(Common.CONFIG.SHOW_EFFECTS()) {
+            world.addParticle(particle, (double)pos.getX() + x, (double)pos.getY() + y, (double)pos.getZ() + z, 0.0D, 0.0D, 0.0D);
         }
     }